+++ docutils/HISTORY.rst
@@ -21,6 +21,12 @@ Release 0.23b0 (unpublished)
 
   - More consistent and concise command line help.
 
+* docutils/parsers/rst/states.py
+
+  - Relax "section title" system message from SEVERE to ERROR.
+  - Ensure new "current node" is valid when switching section level
+    (cf. bugs #508 and #509).
+
 
 Release 0.22 (2026-07-29)
 =========================